🧹 os: fold the new detector tests into detector_platform_test.go
They belong beside the platform tests they extend, not in a file of
their own.
